+++ docs/history.rst
@@ -2,6 +2,13 @@
   * bugfix: cannot bind to IPv6 address :ghissue:`5`.
   * enhancement: Futures waiter_connected, and waiter_closed added to server.
   * change: TelSh.feed_slc merged into TelSh.feed_byte as slc_function keyword.
+  * change: TelnetServer.CONNECT_DEFERED class attribute renamed DEFERRED.
+    Default value changed to 50ms from 100ms.
+  * enhancement: Default TelnetServer.PROMPT_IMMEDIATELY = False ensures prompt
+    is not displayed until negotiation is considered final.  It is no longer
+    "aggressive".
+  * enhancement: TelnetServer.pause_writing and resume_writing callback wired.
+  * enhancement: TelSh.pause_writing and resume_writing methods added.
 
 0.2.4
   * bugfix: pip installation issue :ghissue:`8`.